API: added cache size to the cfg_init() function
this provides ease when allocating the cache buffers for the first time.
cfg_init() checks if the user passes a negative value for cache size.
if so, the default cache size is used.

the first call to cfg_cache_size_set() is made in cfg_parse_buffer(),
but the user can still resize the cache buffer before or after a cfg buffer
is parsed by calling cfg_cache_size_set() directly.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>