lets only auto-call cfg_cache_clear() when parsing a new buffer
cfg_cache_clear() is a function used to clear the cache buffer,
but there is really no need to call it when setting the cache
buffer size from cfg_cache_size_set(). we are only going
to call it when parsing a new buffer.

also now if cfg_cache_size_set() and the cache buffer grows,
lets zero-out the section where the buffer grew and preserve
the lower index values near the top [0 - sz_old_buffer].

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>