cfg_cache_size_set: lets clear the cache buffer with free for size 0
C90 and C99 differ for realloc() a bit so lets check if zero is passed
for size in cfg_cache_size_set() and clear the buffers with free().
if a larger than zero value is passed then lets call realloc().

+
some comment changes in cfg2.c

+
added a note that 'make run' can be used to run the test file

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>