cfg2.c: support escaping any character including '='
adding '\' in front of any character will escape this character.
the same now works for '=' and this technically allows '='
inside both keys and values.

there are many performance improvements, while the entire parse
can be optimized and cleaned up greatly as this code was
written in a couple of days and is not very readable.

the new escape functionality is not tested very well!

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>