add a cfg_char **section field to cfg_t
once the parser traverses the sections and has found
how many there are, we store each name in st->section[idx].

this allows the user to traverse and entry list and match
entries to sections with hash - making it possible for
the user to be aware of the section names without having
access to the file for review. the library does not do
assign automatically entries to sectons as it could be
redundant and requires an extra pass of traversing the
entrire list if we don't want to use a lot of realloc
calls or growing lists.

potentially, if a cfg_section_t object exits we can assign
all related cfg_entry_t pointers to it and then the retriaval
of a cfg_entry_t a from section could be much faster, but again,
this could be a redundant optimization in some cases.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>