make cfg_parse_buffer() accept buffer size again
if the user wants to parse only a section of a larger
constant buffer, the sz argument is obligatory.

because it won't be possible to NULL terminate said buffer
at the exact end location which is required and instead
a new buffer needs to be allocated.

this change does not complicate the inner static methods much.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>