handle missing equal signs on lines which aren't
if the last line doesn't end with '\' this means that
the new line should be a line which has a key.

if an equal sign is not present on that line a warning
is shown and the value on this line is appended at the
begging of the next key.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>