cfg_unescape(): fix an buffer overwrite bug
the way cfg_unescape() works is the following:
- it receive a buffer with a certain size which is the
raw input for the parser
- it goes through it and converts it into a 'packed' buffer
with key, value, sections separated with special characters
- [*] the input buffer is the output buffer, which means that
the parsing is written to the same input buffer

the last step [*] has a potential problem:
- we need to see if the last read character of the input buffer
was \n, for instance.

now oddly enough the input-output buffer being the same works
in most use cases, but it fails if the first few, keys/values
in the input will have the exact same length in the output buffer.
because characters such as \n in the source will be overriden.

to solve the issue a last_char variable is introduced which
stores the last variable before writing the destination character
and completing the for loop.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>