optimize linear search in terms of section matching
so if no optimization flags are added, a bench test
with 2GB of allocated RAM shows that skipping an iteration
with 'continue' if one of the integers (e.g. section hash) does
not match is slightly faster that doing the "match_key && match_section".

enabling -O2 again shows a slight benefit of skipping with
'continue'.

i haven't looked at the genarated assembly at all for this one,
but it's certainly compiler and CPU dependent.

eventually the one huge cfg_entry_t buffer could be separated
into cfg_section_t having their own cfg_entry_t buffers, which
will be a performance boost in terms of reading data, but the
parser will become slower because another iteration over the
whole parser input will have to estimate how many entries per
section are there, given we don't want to call realloc() all
the time. currently realloc() is only called multiple times
in cfg_write_buffer() out of laziness.

so having separate section buffers will make:
- the parser slower
- linear search faster
- delete faster
- the writer faster

might be a vialable option, but the thing is, this is negligable on
a modern CPU as we are talking differences in bare split seconds
for giggabytes of data, which isn't exactly the amount of configuration
one should store.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>