API: remove cfg_entry_t.index
the index property is a bit of a burden at this point.
given each entry is in the same continous memory buffer (cfg_t.entry)
we can easily obtain the position of an entry in the buffer by
subtraction the pointer of the base element &cfg_t.entry[0] from the
pointer of the cfg_entry_t element we want the index of.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>