Makefile/cfg2.h: support __declspec(dllimport) on win32
prevent the eventual thunk generation by using __declspec(dllimport)
when the library is a DLL build. result would be a indirect call in the
import table, instead of the possible "jmp DWORD PTR XX".

this however imposes the requirement for -DCFG_LIB_STATIC for users
if they want to link statically on win32.

also rename CFG_EXPORT to CFG_API for a better context.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>